[Prototype] Fixed an 'Error parsing class' when trying to load the generated Jimple classes.
Problem: the Java bytecode spec allows almost all of unicode in identifiers, but the
Jimple parser is a bit more conservative. But they can be quoted however.

Solution: Add single quotes (double quotes don't work) to everything that could be a type
 or an identifier just to be sure.

Implementation: Added single quotes in the WS wrapper template where it looked appropriate
to do so.